session: Reimplement using sub-commands
add license headers

edit help messages

simplify arg retrieval

add --force flag to 'session clear' command

simplify clear logic

add singular sessionResume

reword comment

%s/useless/obsolete

fix help text

only clear selected sessions

fix imports
